pyproject.toml
[project]
name = "youandme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "faiss-cpu>=1.13.0",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-groq>=1.0.1",
    "python-dotenv>=1.2.1",
    "pytube>=15.0.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "youtube-transcript-api>=1.2.3",
]
